SimpleChatTC:SimpleProxy: Move web requests to its own module
